+++ engine/flink/model-util/src/main/java/pl/touk/nussknacker/engine/flink/util/transformer/aggregate/SessionWindowTrigger.java
@@ -1,0 +1,19 @@
+package pl.touk.nussknacker.engine.flink.util.transformer.aggregate;
+
+public enum SessionWindowTrigger {
+
+    OnEvent("On each event"),
+    OnEnd("After session end");
+
+    private final String label;
+
+    SessionWindowTrigger(String label) {
+        this.label = label;
+    }
+
+    @Override
+    public String toString() {
+        return label;
+    }
+
+}